requirements.txt
flask==3.0.0
flask-cors==4.0.0
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
anthropic==0.34.2
python-dotenv==1.0.0

